Update SlayerOverlay.java (#450)

+++ src/main/java/io/github/moulberry/notenoughupdates/overlays/SlayerOverlay.java
@@ -222,7 +222,7 @@ public class SlayerOverlay extends TextOverlay {
 
 		untilNextSlayerLevel = xpToLevelUp - slayerIntXP;
 		if (xpPerBoss != 0 && untilNextSlayerLevel != 0 && xpToLevelUp != 0) {
-			bossesUntilNextLevel = (int) Math.ceil((xpToLevelUp - untilNextSlayerLevel) / xpPerBoss);
+			bossesUntilNextLevel = (int) Math.ceil((float) (xpToLevelUp - untilNextSlayerLevel) / (float)(xpPerBoss));
 		} else {
 			bossesUntilNextLevel = 0;
 		}
